requirements.txt
mysql-connector-python
pandas
Office365-REST-Python-Client
google-api-python-client
google-cloud-bigquery
openpyxl
